+++ index.js
@@ -37,6 +37,7 @@ import "./features/inventory/showRarityUpgrade";
 import "./features/inventory/highlightAttributeFusionMatchingItems";
 import "./features/inventory/showPricePerT1Shard";
 import "./features/inventory/showExpertiseKills";
+import "./features/inventory/showCaughtTrophyFishRaritiesInOdger";
 
 import "./features/chat/announceMobSpawnToAllChat";
 import "./features/chat/messageOnPlayerDeath";