+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 Released: ???
 
 Features:
+- Added option to compact sea creature catch messages [disabled by default]. It shortens double hook message and catch message that says what sea creature you caught.
 - Added option to share the location to ALL chat on Plhlegblast catch [disabled by default].
 
 Bugfixes: