+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ Released: ???
 Features:
 - Added option to compact sea creature catch messages [disabled by default]. It shortens double hook message and catch message that says what sea creature you caught.
 - Added option to share the location to ALL chat on Plhlegblast catch [disabled by default].
+- Added option to show caught trophy fish rarities in Odger's Trophy Fishing GUI. [disabled by default] 
 
 Bugfixes:
 - Changed Plhlegblast's rarity to legendary in the alert and Rare Catches tracker.